Tutorials (3)

Plotting

Currently, there are two plotting scripts included with Obliqua. More will be added.

Obliqua.plotting.plot_imagk2_spectrum — Method

plot_imagk2_spectrum(σ_range, imag_k2; outpath)

Create and save a plot of the k2 Lovenumber as function of forcing frequency.

Arguments

σ_range::Array{prec,1} : Forcing frequency range.

imag_k2::Array{precc,1} : k2 Lovenumbers.

Keyword Arguments

outpathn::Union{Nothing,String}=nothing : If provided, the figure is saved to this path using savefig.

Returns

plt : The Plots.Plot object.

Obliqua.plotting.save_heat_profile — Method

save_heat_profile(radius, power_prf; filename="heat_profile.png")

Create and save a plot of the heat dissipation profile as a function of radius.

Returns

plt : The Plots.Plot object.